Proyectos/empleados/requirements.txt
Django == 3.2.9
Unipath == 1.1
psycopg2 == 2.9.1
asgiref==3.4.1
pathlib2==2.3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
django-ckeditor==6.1.0

